Refactor: move server to mcp server package
