+++ tsconfig.json
@@ -9,6 +9,6 @@
     "outDir": "./lib"
   },
   "include": [
-    "src", "packages/core/tools/common.ts", "packages/core/tools/screenshot.ts", "packages/core/tools/snapshot.ts", "packages/core/utils.ts", "packages/testickl/endtoend.ts", "packages/core/tools/tool.ts", "packages/mcp/server/index.ts", "packages/core/context.ts",
+    "src", "packages/core/tools/common.ts", "packages/core/tools/screenshot.ts", "packages/core/tools/snapshot.ts", "packages/core/utils.ts", "packages/testickl/endtoend.ts", "packages/core/tools/tool.ts", "packages/mcp/server/index.ts", "packages/core/context.ts", "packages/mcp/server/server.ts",
   ],
 }
